% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centering.R
\name{Centering}
\alias{Centering}
\title{Centering Caption of LaTeX Table}
\usage{
Centering(x)
}
\arguments{
\item{x}{LaTeX-formatted table with a caption.}
}
\value{
LaTeX-formatted table with a centered caption.
}
\description{
This function adds the "\\centering" command to the caption of a LaTeX table,
allowing the caption to be centered in the final output.
}
\examples{
latex_table <- "\\\\begin{table}\\\\caption{This is a caption.}\\\\end{table}"
centered_table <- Centering(latex_table)
print(centered_table)
}
