% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_rtl_header.R
\name{markdown_rtl_header}
\alias{markdown_rtl_header}
\title{Generate a YAML header for an R Markdown document with RTL support}
\usage{
markdown_rtl_header()
}
\value{
A character string containing the YAML header.
}
\description{
This function generates a YAML header for an R Markdown document that will produce a PDF output
with support for Persian typesetting using the \code{xepersian} package and the \code{Arial} font.
}
